backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Flask-Caching==2.1.0
requests==2.31.0
Pillow==10.1.0
